src/M5-Sprint-4/Projeto-Kenzie-Buster/requirements.txt
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
black==23.7.0
click==8.1.6
decorator==5.1.1
Django==4.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
executing==1.2.0
iniconfig==2.0.0
ipython==8.14.0
jedi==0.19.0
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
packaging==23.1
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.10.0
pluggy==1.2.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
PyJWT==2.8.0
pytest==7.4.0
pytest-django==4.5.2
pytest-testdox==3.1.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
wcwidth==0.2.6
